integration: RBAC tests for /resize request
The test confirms that the subject can successfully
resize the Pod resources but not the entire pod.
